Rename maps/base.py to core.py
