+++ gammapy/maps/core.py
